First build of CSV_generator.py file
This file contains functions that create CSV files with different kinds of defects (and one that creates files with no defects), as well as a function that can take the number of each type of file you would like to create for a certain date and call all of the other functions for you. It's also already set up to ask what date and number of each file type you would like when run, which will then be created for you. (For testing)