Added file archiver and function to extract filename from file path
Added a small function in extract_filename.py that  takes a file path as input and returns the filename.
Also created a function in archive_checked_file.py that takes the path of a validated MED_DATA file as input and archives it in the Archive folder, in a folder named after the date on the file.